Player Finalization

+++ Noise.java
@@ -1,14 +1,14 @@
 // JAVA REFERENCE IMPLEMENTATION OF IMPROVED NOISE - COPYRIGHT 2002 KEN PERLIN.
 
 public final class Noise {
-   static public double noise(double x, double y, double z) {
+   static public float noise(float x, float y, float z) {
       int X = (int)Math.floor(x) & 255,                  // FIND UNIT CUBE THAT
           Y = (int)Math.floor(y) & 255,                  // CONTAINS POINT.
           Z = (int)Math.floor(z) & 255;
       x -= Math.floor(x);                                // FIND RELATIVE X,Y,Z
       y -= Math.floor(y);                                // OF POINT IN CUBE.
       z -= Math.floor(z);
-      double u = fade(x),                                // COMPUTE FADE CURVES
+      float u = fade(x),                                // COMPUTE FADE CURVES
              v = fade(y),                                // FOR EACH OF X,Y,Z.
              w = fade(z);
       int A = p[X  ]+Y, AA = p[A]+Z, AB = p[A+1]+Z,      // HASH COORDINATES OF
@@ -23,11 +23,11 @@ public final class Noise {
                              lerp(u, grad(p[AB+1], x  , y-1, z-1 ),
                                      grad(p[BB+1], x-1, y-1, z-1 ))));
    }
-   static double fade(double t) { return t * t * t * (t * (t * 6 - 15) + 10); }
-   static double lerp(double t, double a, double b) { return a + t * (b - a); }
-   static double grad(int hash, double x, double y, double z) {
+   static float fade(float t) { return t * t * t * (t * (t * 6 - 15) + 10); }
+   static float lerp(float t, float a, float b) { return a + t * (b - a); }
+   static float grad(int hash, float x, float y, float z) {
       int h = hash & 15;                      // CONVERT LO 4 BITS OF HASH CODE
-      double u = h<8 ? x : y,                 // INTO 12 GRADIENT DIRECTIONS.
+      float u = h<8 ? x : y,                 // INTO 12 GRADIENT DIRECTIONS.
              v = h<4 ? y : h==12||h==14 ? x : z;
       return ((h&1) == 0 ? u : -u) + ((h&2) == 0 ? v : -v);
    }
